Change redundant folder structure
